T1-168 add search in every page and connect add stuff endpoint

+++ src/components/Modal/DeleteModal/DeleteModal.jsx
@@ -4,11 +4,27 @@ import HeaderModals from "../../../ui/HeaderModals/HeaderModals";
 import styles from "./DeleteModal.module.css";
 import BackgroundModal from "../BackgroundModal/BackgroundModal";
 import ModalButton from "../../../ui/ModalButton/ModalButton";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { setCurrentModal } from "../../../redux/slices/DataSlice";
-import { deleteCategory } from "../../../redux/actions/DataActions";
+import {
+  deleteCategory,
+  deleteMenuPostition,
+  deleteStockProduct,
+} from "../../../redux/actions/DataActions";
 
 const DeleteModal = (props) => {
+  console.log(props.id, "props");
+  const { deleteType } = useSelector((state) => state.data);
+  const deleteTypes = {
+    deletePosition: () => {
+      dispatch(deleteMenuPostition(props.id));
+    },
+    deleteProduct: () => {
+      dispatch(deleteStockProduct(props.id));
+    },
+    deleteBrach: () => {},
+    deleteEmployer: () => {},
+  };
   console.log(props, "props");
   const dispatch = useDispatch();
   function handleClose() {
@@ -16,7 +32,11 @@ const DeleteModal = (props) => {
   }
 
   function handleDeleteCategory() {
-    dispatch(deleteCategory(props));
+    if (deleteType) {
+      deleteTypes[deleteType]();
+    } else {
+      dispatch(deleteCategory(props));
+    }
   }
 
   return (
@@ -24,7 +44,7 @@ const DeleteModal = (props) => {
       <div className={classnames(styles.modal)}>
         <HeaderModals name={"Удаление"} center={true} />
         <h3 className={classnames(styles.title)}>
-          Вы действительно хотите удалить {props} ?
+          Вы действительно хотите удалить {props?.name} ?
         </h3>
         <div className={classnames(styles.button_block, styles.flex)}>
           <ModalButton buttonType={"outlined"} click={handleClose}>
